% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_correlations.R
\name{edge_correlations}
\alias{edge_correlations}
\title{Extracting Multi-Level Edge Correlations (\code{edge_correlations})}
\usage{
edge_correlations(netwrite_output)
}
\arguments{
\item{netwrite_output}{A list object created by \code{netwrite} for a network with multiple unique edge types. This list must contain the \code{system_level_measures} dataframe. If the network stored in \code{netwrite_output} contains only a single edge type, \code{edge_correlations} will return an error message.}
}
\value{
\code{edge_correlations} returns a data frame containing each unique pair of edge types in the network and their respective correlation value.
}
\description{
Given a network with multiple unique edge types, \code{netwrite} calculates the multi-level edge correlations for each unique pair of edge types and stores them in the \code{system_level_measures} dataframe. The \code{edge_correlations} function extracts these values from \code{netwrite}'s output so users can interact with them more directly.
}
\examples{
# Use netwrite on an edgelist
nw_flor <- netwrite(nodelist = florentine_nodes,
                    node_id = "id",
                    i_elements = florentine_edges$source,
                    j_elements = florentine_edges$target,
                    type = florentine_edges$type,
                    directed = FALSE,
                    net_name = "florentine")

edge_correlations(nw_flor)
}
