% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_alter_measures.R
\name{ego_alter_measures}
\alias{ego_alter_measures}
\title{Additional Measure Calculation for Ego Networks (\code{ego_alter_measures})}
\usage{
ego_alter_measures(
  ego_df,
  alter_df,
  ego_prefix = NULL,
  ego_suffix = NULL,
  alter_prefix = NULL,
  alter_suffix = NULL,
  vars = NULL,
  measures = c("diversity", "number_homophilous", "proportion_homophilous", "ei_index",
    "pearson_phi", "euclidean_distance"),
  na.rm = FALSE
)
}
\arguments{
\item{ego_df}{A data frame containing measures of ego attributes. If using data objects created by \code{\link{ego_netwrite}}, this should be the data frame entitled \code{egos}.}

\item{alter_df}{A data frame containing measures of alter attributes. If using data objects created by \code{\link{ego_netwrite}}, this should be the data frame entitled \code{alters}.}

\item{ego_prefix}{A character value indicating a common prefix for variables contained in \code{ego_df}. These prefixes are removed in order to match variables with their counterparts in \code{alter_df}.}

\item{ego_suffix}{A character value indicating a common suffix for variables contained in \code{ego_df}. These suffixes are removed in order to match variables with their counterparts in \code{alter_df}.}

\item{alter_prefix}{A character value indicating a common prefix for variables contained in \code{alter_df}. These prefixes are removed in order to match variables with their counterparts in \code{ego_df}.}

\item{alter_suffix}{A character value indicating a common suffix for variables contained in \code{alter_df}. These suffixes are removed in order to match variables with their counterparts in \code{ego_df}.}

\item{vars}{A character vector indicating which variables appearing in both \code{ego_df} and \code{alter_df} should be matched and used for calculating measures. If the user does not manually specify variables in this argument, \code{ego_alter_measures} will automatically search for matching variables across \code{ego_df} and \code{alter_df}.}

\item{measures}{A character vector indicating which measures should be calculated. Measures include diversity, number and proportion of homophilous alters, E-I index, Pearson's phi, and Euclidean distance. If the user does not manually specify measures in this argument, \code{ego_alter_measures} will calculate all applicable measures to each variable.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be excluded when calculating continuous measures.}
}
\value{
\code{ego_alter_measures} returns a data frame containing the measures specified by the user for each individual ego network. This data frame is complementary to the \code{summaries} data frame created by \code{\link{ego_netwrite}}, and the two can easily be merged.
}
\description{
Common measures of ego network properties can differ in whether they require information about ego's attributes. Some measures, like the h-index, only require information about alter attributes, while others, such as Pearson's phi, require information about both ego and their alters. `ego_netwrite` calculates measures of the former sort by default. `ego_alter_measures`, by contrast, allows users to calculate measures that require both ego- and alter-level data.
}
