% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{florentine_edges}
\alias{florentine_edges}
\title{Edgelist of marriage alliances and business relationships between Florentine families during the Italian Renaissance}
\format{
A data frame with 35 rows and 4 columns:
\describe{
  \item{source}{Outgoing node}
  \item{target}{Receiving node}
  \item{weight}{A placeholder variable for tie/edge weights, set to 1}
  \item{type}{Relation type}
  ...
}
}
\source{
John Padgett
}
\usage{
florentine_edges
}
\description{
Breiger & Pattison (1986), in their discussion of local role analysis, use a subset of data on
the social relations among Renaissance Florentine families collected by John Padgett
from historical documents. The two relations are business ties (recorded financial ties such as loans,
credits and joint partnerships) and marriage alliances. This dataset has since become a standard for illustrating
role analysis methods and working with networks featuring multiple types of relations.
}
\references{
Ronald L. Breiger and Philippa E Pattison. 1986. "Cumulated social roles: The duality of persons and their algebras." \emph{Social Networks} 8(13):215-256.
}
\keyword{datasets}
