% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{florentine_nodes}
\alias{florentine_nodes}
\title{Nodelist of marriage alliances and business relationships between Florentine families during the Italian Renaissance}
\format{
A data frame with 16 rows and 2 columns:
\describe{
  \item{id}{Unique node ID number}
  \item{family}{Name of family corresponding to node}
  ...
}
}
\source{
John Padgett (\href{http://www.casos.cs.cmu.edu/computational_tools/datasets/external/padgett/index2.html}{Website})
}
\usage{
florentine_nodes
}
\description{
Breiger & Pattison (1986), in their discussion of local role analysis, use a subset of data on
the social relations among Renaissance Florentine families collected by John Padgett
from historical documents. The two relations are business ties (recorded financial ties such as loans,
credits and joint partnerships) and marriage alliances. This dataset has since become a standard for illustrating
role analysis methods and working with networks featuring multiple types of relations.
}
\references{
Ronald L. Breiger and Philippa E Pattison. 1986. "Cumulated social roles: The duality of persons and their algebras." \emph{Social Networks} 8(13):215-256.
}
\keyword{datasets}
