% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hightech}
\alias{hightech}
\title{Multiplex Network of Relationships Between Managers of a High-Tech Company}
\format{
A data frame with 312 rows and 4 columns:
\describe{
  \item{node}{Outgoing node}
  \item{target}{Receiving node}
  \item{weight}{A placeholder variable for tie/edge weights, set to 1}
  \item{layer}{Relation type}
  ...
}
}
\source{
Carnegie Mellon University
}
\usage{
hightech
}
\description{
A network of a small hi-tech computer firm that sold, installed, and maintained computer
systems, represented as an edgelist. Relationships in the network can take on three modes:
1 represents advice relationships, 2 represents friendship relationships, and
3 represents chain of command (e.g., "reporting-to").
}
\references{
David Krackhardt. 1987. "Cognitive social structures". \emph{Social Networks} 9(2):104-134. https://doi.org/10.1016/0378-8733(87)90009-8
}
\keyword{datasets}
