% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ideanetViz.R
\name{install_ideanetViz}
\alias{install_ideanetViz}
\title{Install IDEANet Visualization GUI (\code{install_ideanetViz})}
\usage{
install_ideanetViz()
}
\description{
The IDEANet suite includes a Shiny-based GUI designed to help people process their data and perform basic analyses. \code{ideanetViz} also allows users to easily customize network visualizations and access more advanced analysis modules within its GUI.
}
\examples{
\dontrun{
# Run `install_ideanetViz`
install_ideanetViz()

# Open the ideanetViz GUI
library(ideanetViz)
ideanet_viz()
}
}
