% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqv.R
\name{iqv}
\alias{iqv}
\title{Index of Qualitative Variation (\code{iqv})}
\usage{
iqv(ego_id, measure, prefix = NULL, suffix = NULL)
}
\arguments{
\item{ego_id}{A vector of ego identifiers located in an alter dataframe. If using data objects created by \code{ego_netwrite}, this should be the data frame entitled \code{alters}.}

\item{measure}{A vector of alter attributes for a given categorical measure.}

\item{prefix}{A character value indicating the desired prefix for the calculated homophily measure.}

\item{suffix}{A character value indicating the desired suffix for the calculated homophily measure.}
}
\value{
\code{iqv} returns a dataframe of vectors that include the ego identifier and iqv value of diversity for the desired categorical attribute.
}
\description{
A normalized value of the h-index for measuring the diversity of an ego's network for categorical attributes (Perry et al. 2018)
}
\examples{

data(package = "egor", "egos32")
data(package = "egor", "alters32")

iqv(ego_id = alters32$.EGOID, measure = alters32$country)
}
