% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_level_measures.R
\name{largest_weak_component_igraph}
\alias{largest_weak_component_igraph}
\title{Personal greeting}
\usage{
largest_weak_component_igraph(g)
}
\arguments{
\item{data_type}{Your name (character string; e.g. "john doe").}

\item{package}{# Might need to remove this}

\item{message}{}
}
\value{
A character string, capitalized to title case.
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
hello("james bond")
}
