% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marvel}
\alias{marvel}
\title{Character Relations in Marvel Comics}
\format{
A data frame with 9891 rows and 3 columns:
\describe{
  \item{Source}{Outgoing node}
  \item{Target}{Receiving node}
  \item{Weight}{Edge weight}
  ...
}
}
\source{
Melanie Walsh (\href{https://github.com/melaniewalsh/sample-social-network-datasets/tree/master/sample-datasets/marvel}{Github}),
adapted from data originally compiled by Cesc Rosselló, Ricardo Alberich,
and Joe Miro from Russ Chappell (\href{https://www.chronologyproject.com}{Website})
}
\usage{
marvel
}
\description{
A network, represented as edgelist, containing weighted edges between Marvel Comics characters.
Edge weights were calculated based on how many times two characters' names appeared within 15 words
of one another in a comic.
}
\keyword{datasets}
