% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_splitter.R
\name{net_splitter}
\alias{net_splitter}
\title{Network Splitting}
\usage{
net_splitter(i_elements = NULL, j_elements = NULL, weights = NULL, type = NULL)
}
\arguments{
\item{type}{}
}
\value{
A character string, capitalized to title case.
}
\description{
This function is a prerequisite for using `netwrite` on multi-relational networks. It takes as arguments the standard vectors of ego IDs, alter IDs, edge weights, and relation types for edgelists.
}
\examples{
hello("james bond")
}
