% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_level_measures.R
\name{node_level_igraph}
\alias{node_level_igraph}
\title{Personal greeting}
\usage{
node_level_igraph(nodes, g, directed, message)
}
\arguments{
\item{nodes}{Your name (character string; e.g. "john doe").}

\item{message}{}
}
\value{
A character string, capitalized to title case.
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
hello("james bond")
}
