% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_phi.R
\name{pearson_phi}
\alias{pearson_phi}
\title{Pearson's Phi (`pearson_phi`)}
\usage{
pearson_phi(
  ego_id,
  ego_measure,
  alter_ego,
  alter_measure,
  prefix = NULL,
  suffix = NULL
)
}
\arguments{
\item{ego_id}{A vector of unique ego identifiers located in an ego dataframe. If using data objects created by `ego_netwrite`, this should be the data frame entitled `egos`.}

\item{ego_measure}{A vector of attributes corresponding to each ego.}

\item{alter_ego}{A vector of ego identifiers located in an alter dataframe. If using data objects created by `ego_netwrite`, this should be the data frame entitled `alters`.}

\item{alter_measure}{A vector of attributes corresponding to each alter}

\item{prefix}{A character value indicating the desired prefix for the calculated homophily measure.}

\item{suffix}{A character value indicating the desired suffix for the calculated homophily measure.}
}
\value{
`pearson_phi` returns a dataframe of vectors that include the ego identifier and phi value of homophilous preference.
}
\description{
The `pearson_phi` function identifies the underlying homophilous preference of ego based on the distribution of alter attributes in the population (Perry et al. 2018)
}
\examples{

data(package = "egor", "egos32")
data(package = "egor", "alters32")

pearson_phi(ego_id = egos32$.EGOID, ego_measure = egos32$country,
alter_ego = alters32$.EGOID, alter_measure = alters32$country)
}
