% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qap_run.R
\name{qap_run}
\alias{qap_run}
\title{Quadratic Assignment Procedure (`qap_run`).}
\usage{
qap_run(net, dependent = NULL, variables, directed = F, family = "linear")
}
\arguments{
\item{net}{An `igraph` or `network` object.}

\item{dependent}{A string naming the dependent variable of interest.}

\item{variables}{A vector of strings naming the independent variables of interest.}

\item{directed}{A logical statement identifying if the network should be treated as directed. Defaults to False.}

\item{family}{A string identifying the functional form. Options are "linear" and "binomial". Defauts to "linear".}
}
\value{
`qap_run` returns a list of elements `model_results` that include: 

- `covs_df`, a data frame containing term labels, estimates, standard errors and p-values

- `mods_df`, a data frame containing model-level information including the number of observations, AIC and BIC statistics.
}
\description{
The `qap_run` function is a wrapper around `sna`'s Quadratic Assignment Procedure models `netlm()` and `netlogit()`. It expects a networks objects containing dependent and independent variables of interest (e.g, the output from `qap_setup`).
}
\examples{
qap_setup(net = igraph_object,
         variables = "attribute_1",
         methods = c("attribute_2", "attribute_3", "attribute_4"),
         directed = F,
         family = "linear")
}
