% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{triad}
\alias{triad}
\title{A Small Network Containing all Triads and Motifs}
\format{
A matrix with 9 rows and 9 columns
}
\usage{
triad
}
\description{
An adjacency matrix representing a network of 9 nodes, the ties between which form all possible
triads and 3-node motifs that can appear in a directed network.
}
\keyword{datasets}
