% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netwrite_ego.R
\name{ego_netwrite}
\alias{ego_netwrite}
\title{Ego Network Cleaning and Measure Calculation (\code{ego_netwrite})}
\usage{
ego_netwrite(
  egos,
  ego_id,
  alters = NULL,
  alter_id = NULL,
  alter_ego = NULL,
  alter_type = NULL,
  alter_types = NULL,
  max_alters = Inf,
  alter_alter = NULL,
  aa_ego = NULL,
  i_elements = NULL,
  j_elements = NULL,
  directed = FALSE,
  aa_type = NULL,
  missing_code = 99999,
  na.rm = FALSE,
  egor = FALSE,
  egor_design = NULL
)
}
\arguments{
\item{egos}{A data frame containing measures of ego attributes.}

\item{ego_id}{A vector of unique identifiers corresponding to each ego, or a single character value indicating the name of the column in \code{egos} containing ego identifiers.}

\item{alters}{A data frame containing measures of alter attributes.}

\item{alter_id}{A vector of identifiers indicating which alter is associated with a given row in \code{alters}, or a single character value indicating the name of the column in \code{alters} containing alter identifiers.}

\item{alter_ego}{A vector of identifiers indicating which ego is associated with a given alter, or a single character value indicating the name of the column in \code{alters} containing ego identifiers.}

\item{alter_type}{(DEPRECATED) A numeric or character vector indicating the types of relationships existing between ego and a given alter, or a single character value indicating the name of the column in \code{alters} containing relationship type. If \code{alter_type} is specified, \code{ego_netwrite} will treat the data as a set of multi-relational networks and produce additional outputs reflecting the different types of ties occurring in each ego network.}

\item{alter_types}{A character vector indicating the columns in \code{alters} that indicate whether a given alter has certain types of relations with ego. These columns should all contain binary measures indicating whether alter has a particular type of relation with ego.}

\item{max_alters}{A numeric value indicating the maximum number of alters an ego in the dataset could have nominated}

\item{alter_alter}{A data frame containing an edgelist indicating ties between alters in each ego's network. This edgelist is optional, but \code{ego_netwrite} will not provide certain measures without it.}

\item{aa_ego}{A vector of identifiers indicating which ego is associated with a given tie between alters, or a single character indicating the name of the column in \code{alter_alter} containing ego identifiers.}

\item{i_elements}{A vector of identifiers indicating which alter is on one end of an alter-alter tie, or a single character indicating the name of the column in \code{alter_alter} containing these identifiers.}

\item{j_elements}{A vector of identifiers indicating which alter is on the other end of an alter-alter tie, or a single character indicating the name of the column in \code{alter_alter} containing these identifiers.}

\item{directed}{A logical value indicating whether network ties are directed or undirected.}

\item{aa_type}{A numeric or character vector indicating the types of relationships represented in the alter edgelist, or a single character value indicating the name of the column in \code{alter_alter} containing relationship type. If \code{alter_type} is specified, \code{ego_netwrite} will treat the data as a set of multi-relational networks and produce additional outputs reflecting the different types of ties occurring in each ego network.}

\item{missing_code}{A numeric value indicating "missing" values in the alter-alter edgelist.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be excluded when calculating continuous measures.}

\item{egor}{A logical value indicating whether output should include an \code{egor} object, which is often useful for visualizaton and for simulation larger networks from egocentric data.}

\item{egor_design}{If creating an \code{egor} object, a list of arguments to \code{\link[srvyr:as_survey_design]{srvyr::as_survey_design}} specifying the sampling design for egos. This argument corresponds to \code{ego_design} in \code{\link[egor:egor]{egor::egor}}.}
}
\value{
\code{ego_netwrite} returns a list containing several output objects. Users may find it easier to access and work with outputs by applying \link{list2env} to this list, which will separate outputs and store them in the R Global Environment. Outputs include a data frame containing measures of ego attributes, another data frame containing measures of alter attributes and network position, a third containing the alter-alter edgelist (when applicable), a fourth containing summary measures for each individual ego network, and a fifth providing summary measures for the overall dataset. Additionally, \code{ego_netwrite} returns a list of \code{igraph} objects constructed for each individual ego network, as well as an \code{egor} object for the overall dataset if desired.
}
\description{
The \code{ego_netwrite} function reads in data pertaining to ego networks and processes them into a set of standardized outputs, including measures commonly calculated for ego networks.
}
\examples{

egonets <- ego_netwrite(egos = egor::egos32,
                        ego_id = ".EGOID",

                        alters = egor::alters32,
                        alter_id = ".ALTID",
                        alter_ego = ".EGOID",

                        alter_alter = egor::aaties32,
                        aa_ego = ".EGOID",
                        i_elements = ".SRCID",
                        j_elements = ".TGTID",

                        missing_code = NULL,
                        directed = FALSE,

                        egor = TRUE)

list2env(egonets, .GlobalEnv)

# View summaries of individual ego networks
head(summaries)

# View summary of overall dataset
head(overall_summary)

# View sociogram of second ego network
plot(igraph_objects[[2]]$igraph_ego)
}
