% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{football}
\alias{football}
\title{American College Football}
\format{
An `igraph` object containing 613 edges between 115 vertices (nodes). Vertices contain three attributes:
\describe{
  \item{id}{Unique identification number.}
  \item{label}{Name of college team represented by vertex.}
  \item{value}{A numberic indicator of football conference affiliation.}
  ...
}
}
\source{
Included by permission of M. Girvan and M. E. J. Newman (\href{https://websites.umich.edu/~mejn/netdata/}{Website})
}
\usage{
football
}
\description{
Network of American football games between Division IA colleges during regular season Fall 2000.
}
\references{
M. Girvan and M. E. J. Newman. 2002. "Community structure in social and biological networks." \emph{Proc. Natl. Acad. Sci. USA} 99:7821-7826.
}
\keyword{datasets}
